refactor(frontend): extract ChangeAuthTokenForm to a separate file
